TO DO: - check links - probably all broken
-------------------------------------------------
Links problems have to do with .htaccess.
Need to find a way to access resources
without making everything public.
Create a routing class which is self-aware
of .htaccess root taking as argument a path
relative to the page where that route was
called from and return a path relative
to the site root directory.
-------------------------------------------------

- redo all the css into a global css file
- redo all the js into a global js file
- ...